pyproject.toml
[project]
name = "distributed-data-storage"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "numpy>=2.4.0",
    "pandas>=2.3.3",
    "psycopg[binary,pool]>=3.3.2",
    "pyarrow>=22.0.0",
    "python-dotenv>=1.2.1",
    "s3fs>=2025.12.0",
    "sqlalchemy>=2.0.45",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
]
